tweak: move isGenesisIDMissing, isApiMissing to walletValidation and add spec to wallet tests
